+++ part3/app/services/facade.py
@@ -1,4 +1,5 @@
-from app.persistence.repository import InMemoryRepository
+from app.persistence.user_repository import UserRepository
+from app.persistence.repository import SQLAlchemyRepository
 from app.models.user import User
 from app.models.amenity import Amenity
 from app.models.place import Place
@@ -6,17 +7,19 @@ from app.models.review import Review
 
 class HBnBFacade:
     def __init__(self):
-        self.user_repo = InMemoryRepository()
-        self.amenity_repo = InMemoryRepository()
-        self.place_repo = InMemoryRepository()
-        self.review_repo = InMemoryRepository()
+        # Use SQLAlchemyRepository for all entities
+        self.user_repo = UserRepository()
+        self.amenity_repo = SQLAlchemyRepository(Amenity)
+        self.place_repo = SQLAlchemyRepository(Place)
+        self.review_repo = SQLAlchemyRepository(Review)
 
-    # USER
+    # ------------------- USER -------------------
     def create_user(self, user_data):
         user = User(**user_data)
+        user.hash_password(user_data['password'])
         self.user_repo.add(user)
         return user
-    
+
     def get_users(self):
         return self.user_repo.get_all()
 
@@ -25,11 +28,11 @@ class HBnBFacade:
 
     def get_user_by_email(self, email):
         return self.user_repo.get_by_attribute('email', email)
-    
+
     def update_user(self, user_id, user_data):
         self.user_repo.update(user_id, user_data)
-    
-    # AMENITY
+
+    # ------------------- AMENITY -------------------
     def create_amenity(self, amenity_data):
         amenity = Amenity(**amenity_data)
         self.amenity_repo.add(amenity)
@@ -44,25 +47,24 @@ class HBnBFacade:
     def update_amenity(self, amenity_id, amenity_data):
         self.amenity_repo.update(amenity_id, amenity_data)
 
-    # PLACE
+    # ------------------- PLACE -------------------
     def create_place(self, place_data):
-        user = self.user_repo.get_by_attribute('id', place_data['owner_id'])
-        if not user:
+        owner = self.user_repo.get(place_data['owner_id'])
+        if not owner:
             raise KeyError('Invalid input data')
-        del place_data['owner_id']
-        place_data['owner'] = user
+
         amenities = place_data.pop('amenities', None)
+        place_data['owner'] = owner
+        place = Place(**place_data)
+
         if amenities:
             for a in amenities:
                 amenity = self.get_amenity(a['id'])
                 if not amenity:
                     raise KeyError('Invalid input data')
-        place = Place(**place_data)
-        self.place_repo.add(place)
-        user.add_place(place)
-        if amenities:
-            for amenity in amenities:
                 place.add_amenity(amenity)
+
+        self.place_repo.add(place)
         return place
 
     def get_place(self, place_id):
@@ -74,26 +76,19 @@ class HBnBFacade:
     def update_place(self, place_id, place_data):
         self.place_repo.update(place_id, place_data)
 
-    # REVIEWS
+    # ------------------- REVIEW -------------------
     def create_review(self, review_data):
         user = self.user_repo.get(review_data['user_id'])
-        if not user:
-            raise KeyError('Invalid input data')
-        del review_data['user_id']
-        review_data['user'] = user
-        
         place = self.place_repo.get(review_data['place_id'])
-        if not place:
+        if not user or not place:
             raise KeyError('Invalid input data')
-        del review_data['place_id']
-        review_data['place'] = place
 
+        review_data['user'] = user
+        review_data['place'] = place
         review = Review(**review_data)
         self.review_repo.add(review)
-        user.add_review(review)
-        place.add_review(review)
         return review
-        
+
     def get_review(self, review_id):
         return self.review_repo.get(review_id)
 
@@ -111,10 +106,5 @@ class HBnBFacade:
 
     def delete_review(self, review_id):
         review = self.review_repo.get(review_id)
-        
-        user = self.user_repo.get(review.user.id)
-        place = self.place_repo.get(review.place.id)
-
-        user.delete_review(review)
-        place.delete_review(review)
-        self.review_repo.delete(review_id)
+        if review:
+            self.review_repo.delete(review_id)
